requirements.txt
dj-database-url==0.5.0
Django==2.2.6
djangorestframework==3.10.3
gunicorn==20.0.4
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
psycopg2==2.6.2